+++ PlayGen.SUGAR.Unity/PlayGen.SUGAR.Unity/Group/BaseGroupMemberInterface.cs
@@ -56,7 +56,7 @@ namespace PlayGen.SUGAR.Unity
 		/// </summary>
 		protected override string LoadErrorText()
 		{
-			return Localization.Get("GROUPS_LOAD_ERROR");
+			return Localization.Get("GROUPS_MEMBER_LOAD_ERROR");
 		}
 
 		/// <summary>