+++ CHANGELOG.md
@@ -10,7 +10,9 @@ The following will be available in the next version of SUGAR Unity:
 - Update Callbacks now triggered when calling GameData methods with no signed in user.
 - Update Auto Log-in Editor tool values now saved per project (according to the Product Name) on the same machine.
 - Update interfaces and unity clients related to relationships in order to better fit expected functionality and to make them more consistent with the rest of the project.
+- Update to provided localization resource.
 - Fix for being locked into UI when changing aspect ratio from landscape to portrait and vice versa.
+- Fix for streaming assets config file being required in order to use the seeding functionality.
 
 ### 1.2.2
 - Update SUGAR url.